renamed gráficos->GUI, finalizado max-min, añadido wxChoice para más algoritmos.

+++ maxmin.h
@@ -13,7 +13,7 @@ public:
     static bool verbo;
     static double umbral;
     static double dist_mayor;
-    static std::vector<std::vector<float>> matrizDistancias;
+    static std::vector<std::vector<double>> matrizDistancias;
     static void max_min_ini(wxTextCtrl *out);
     static void realizarClasificacion(wxTextCtrl *out);
     static std::string logM(int pos);
